requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
umap-learn
shap
mofapy2
pyyaml